refactor: reorganize message types to avoid deep-nesting typescript warning
