+++ package/src/components/Message/MessageSimple/MessageDeleted.tsx
@@ -21,8 +21,7 @@ import {
 
 import type { MessageFooterProps } from './MessageFooter';
 
-import type { MessageType } from '../../MessageList/hooks/useMessageList';
-
+import type { MessageType } from '../../../types/messageTypes';
 import type {
   DefaultAttachmentType,
   DefaultChannelType,
