+++ package/src/components/Message/MessageSimple/utils/renderText.tsx
@@ -13,8 +13,7 @@ import {
   ReactNodeOutput,
 } from 'simple-markdown';
 
-import type { MessageType } from '../../../MessageList/hooks/useMessageList';
-
+import type { MessageType } from '../../../../types/messageTypes';
 import type {
   DefaultAttachmentType,
   DefaultChannelType,
