+++ package/src/components/Message/utils/removeReservedFields.ts
@@ -1,4 +1,4 @@
-import type { MessageType } from '../../MessageList/hooks/useMessageList';
+import type { MessageType } from '../../../types/messageTypes';
 import type {
   DefaultAttachmentType,
   DefaultChannelType,
