+++ package/src/contexts/imageGalleryContext/ImageGalleryContext.tsx
@@ -2,7 +2,7 @@ import React, { PropsWithChildren, useContext, useState } from 'react';
 
 import { getDisplayName } from '../utils/getDisplayName';
 
-import type { MessageType } from '../../components/MessageList/hooks/useMessageList';
+import type { MessageType } from '../../types/messageTypes';
 import type {
   DefaultAttachmentType,
   DefaultChannelType,
