+++ package/src/types/messageTypes.ts
@@ -1,0 +1,38 @@
+import type { ChannelState, MessageResponse } from 'stream-chat';
+import type {
+  DefaultAttachmentType,
+  DefaultChannelType,
+  DefaultCommandType,
+  DefaultEventType,
+  DefaultMessageType,
+  DefaultReactionType,
+  DefaultUserType,
+  UnknownType,
+} from './types';
+
+export type GroupType = 'bottom' | 'middle' | 'single' | 'top';
+
+export type MessagesWithStylesReadByAndDateSeparator<
+  At extends UnknownType = DefaultAttachmentType,
+  Ch extends UnknownType = DefaultChannelType,
+  Co extends string = DefaultCommandType,
+  Me extends UnknownType = DefaultMessageType,
+  Re extends UnknownType = DefaultReactionType,
+  Us extends UnknownType = DefaultUserType,
+> = MessageResponse<At, Ch, Co, Me, Re, Us> & {
+  groupStyles: GroupType[];
+  readBy: boolean | number;
+  dateSeparator?: Date;
+};
+
+export type MessageType<
+  At extends UnknownType = DefaultAttachmentType,
+  Ch extends UnknownType = DefaultChannelType,
+  Co extends string = DefaultCommandType,
+  Ev extends UnknownType = DefaultEventType,
+  Me extends UnknownType = DefaultMessageType,
+  Re extends UnknownType = DefaultReactionType,
+  Us extends UnknownType = DefaultUserType,
+> =
+  | ReturnType<ChannelState<At, Ch, Co, Ev, Me, Re, Us>['formatMessage']>
+  | MessagesWithStylesReadByAndDateSeparator<At, Ch, Co, Me, Re, Us>;
